+++ test/vagrant/ssh_test.rb
@@ -195,16 +195,15 @@ class SshTest < Test::Unit::TestCase
     end
 
     should "return false if SSH connection times out" do
-      Net::SSH.expects(:start)
-      assert !@ssh.up?
-    end
+      @env.config.ssh.timeout = 0.5
 
-    should "allow the thread the configured timeout time" do
-      @thread = mock("thread")
-      @thread.stubs(:[])
-      Thread.expects(:new).returns(@thread)
-      @thread.expects(:join).with(@env.config.ssh.timeout).once
-      @ssh.up?
+      Net::SSH.stubs(:start).with() do
+        # Sleep here to artificially fake timeout
+        sleep 1
+        true
+      end
+
+      assert !@ssh.up?
     end
 
     should "return false if the connection is refused" do